Separate WAFS verification from global_det. - Inside code, change to COMPONENT=wafs and RUN=atmos - script files are renamed - WAFS verification has its own vertical structure:   - ecf/wafs   - jobs/wafs   - scripts/wafs   - ush/wafs   - parm/metplus_config/wafs   - parm/evs_config/wafs
Speed up stat runtime by adding more MPMD processes
- On top of observation data (GFS GCIP), run at resolutions (0P25 1P25) parallel and centers (blend UK US) parallel for icing.
- Increase number of CPUs from 2 to 7
- Runtime is down to 10 minutes

Speed up plotting runtime by adding more MPMD processes
- Run on different domains (9 subdomains + 1 global) parallel instead of sequentially
- Multiplied the number of CPUs by the number of domains, increased from 4 to 40.
- Runtime is down from 40 minutes to 16

Add email alerts when observation data are missing

Modify Python code to plot a domain at one time. Global (all) is a domain which collects data of all subdomains and is plotted once.
