+++ ecf/wafs/stats/jevs_wafs_atmos_stats.ecf
@@ -1,12 +1,17 @@
-#PBS -N jevs_aviation_plots
+#PBS -N jevs_wafs_stats
 #PBS -j oe
 #PBS -S /bin/bash
 #PBS -q dev
 #PBS -A VERF-DEV
-#PBS -l walltime=06:00:00
-#PBS -l select=1:ncpus=4:mem=200GB
+#PBS -l walltime=00:30:00
+#PBS -l select=1:ncpus=7:mem=8GB
 #PBS -l debug=true
 #PBS -V
+
+##PBS -q debug
+##PBS -A GFS-DEV
+##PBS -q dev
+##PBS -A VERF-DEV
 
 export OMP_NUM_THREADS=1
 
@@ -36,6 +41,7 @@ module load cfp/$cfp_ver
 module load gsl/$gsl_ver
 module load met/$met_ver
 module load metplus/$metplus_ver
+export MET_bin_exec=bin
 
 set -xa
 module list
@@ -43,36 +49,38 @@ module list
 ############################################################
 # environment variables set
 ############################################################
-export envir=dev
+export envir=prod
 
 export NET=evs
-export STEP=plots
-export COMPONENT=global_det
-export RUN=aviation
+export STEP=stats
+export COMPONENT=wafs
+export RUN=atmos
 export VERIF_CASE=grid2grid
-
-export COMIN=/lfs/h2/emc/vpppg/noscrub/$USER/evs/$evs_ver
 
 export FIXevs=/lfs/h2/emc/vpppg/noscrub/emc.vpppg/verification/EVS_fix
 
-export USH_DIR=$HOMEevs/ush/$COMPONENT
-export DAYS_LIST="90 31"
+export COMOUT=/lfs/h2/emc/vpppg/noscrub/$USER/evs/$evs_ver
+
+############################################################
+# set up for email alerts of missing data
+############################################################
+export pid=${PBS_JOBID:-$$}
+export job=${PBS_JOBNAME:-jevs_aviation_stats}
+export jobid=$job.$pid
+export maillist=${maillist:-'geoffrey.manikin@noaa.gov,yali.mao@noaa.gov'}
 
 ############################################################
 # CALL executable job script here
 ############################################################
-export KEEPDATA=YES
-
-export pid=${pid:-$$}
 export DATA=/lfs/h2/emc/ptmp/$USER/evs/working/${STEP}.$pid
 
-$HOMEevs/jobs/global_det/plots/JEVS_GLOBAL_DET_AVIATION_PLOTS
+export KEEPDATA=YES
+
+$HOMEevs/jobs/wafs/stats/JEVS_WAFS_ATMOS_STATS
 
 ############################################################
 ## Purpose: This job generates the grid2grid statistics stat
 ##          files for GFS WAFS
 ############################################################
 #
-
 exit
-
