+++ parm/evs_config/wafs/config.evs.plots.wafs.atmos
@@ -76,7 +76,7 @@
 set -x
 
 # Case type settings are used to check if settings are allowed. **
-export VERIF_CASE=$RUN
+export VERIF_CASE=$COMPONENT
 export VERIF_TYPE="analysis"
 
 # Used to name the graphics ouput files. File names will start with the URL_HEADER, followed by plot 
@@ -134,7 +134,7 @@ export FCST_VALID_HOUR=$FCST_VALID_HOUR
 export FCST_LEVEL="$VAR1_LEVELS_FCST"
 export OBS_LEVEL="$VAR1_LEVELS_OBS"
 export var_name="$VAR_NAME"
-export VX_MASK_LIST="ASIA, AUNZ, EAST, NAMER, NATL_AR2, NHEM, NPO, SHEM, TROPICS"
+export VX_MASK_LIST=${VX_MASK_LIST:-"ASIA, AUNZ, EAST, NAMER, NATL_AR2, NHEM, NPO, SHEM, TROPICS"}
 
 # Will use statistics for all comma-separated lead times 
 export FCST_LEAD="$FHOURS_EVSlist"
@@ -166,6 +166,6 @@ export COMPONENT=$COMPONENT
 # "stat_by_level", "threshold_average", "time_series", and "valid_hour_average".
 export PLOT_TYPE=$PLOT_TYPE
 # Executes the desired python script.  No need to edit this. 
-python $USHevs/evs_global_det_aviation_plot_${PLOT_TYPE}.py
+python $USHevs/evs_wafs_atmos_plot_${PLOT_TYPE}.py
 
 exit 0
