+++ parm/evs_config/wafs/config.evs.stats.wafs.atmos.standalone
