--- parm/metplus_config/global_det/aviation/plots/plots.conf
@@ -1,1 +1,0 @@
-sample
