--- parm/metplus_config/global_det/aviation/stats/tool_fcstFCST_obsOBS.conf
@@ -1,1 +1,0 @@
-sample
