+++ parm/metplus_config/wafs/atmos/stats/GridStat_fcstWAFS_obsGFS.conf
