+++ parm/metplus_config/wafs/atmos/stats/StatAnalysis_fcstWAFS_obsGCIP_GatherbyDay.conf
