+++ parm/metplus_config/wafs/atmos/stats/StatAnalysis_fcstWAFS_obsGFS_GatherbyDay.conf
