+++ parm/metplus_config/wafs/atmos/stats/StatAnalysis_fcstWAFS_obsGFSwdir_GatherbyDay.conf
