+++ ush/wafs/evs_wafs_atmos_plot_check_variables.py
