+++ ush/wafs/evs_wafs_atmos_plot_settings.py
@@ -735,7 +735,7 @@ class Reference():
                                   'SHEM': 'Southern Hemisphere',
                                   'TROPICS': 'Tropics',
                                   'G193': 'Global',
-                                  'G45': 'Global',
+                                  'G045': 'Global',
                                   'G002': 'Global',
                                   'G003': 'Global',
                                   'G130': 'CONUS - NCEP Grid 130',
@@ -2069,13 +2069,13 @@ class Reference():
                     }
                 }
             },
-            'aviation_analysis': {
+            'wafs_analysis': {
                 'CTC': {
                     'plot_stats_list': ('bias, ets, fss, csi, fbias, fbar,'
                                         + ' obar, pod, farate, faratio, sratio'),
                     'interp': 'NEAREST',
                     'vx_mask_list' : [
-                        'G45','G193','ASIA','AUNZ','EAST','NAMER','NATL_AR2','NHEM','NPO','SHEM','TROPICS'
+                        'G045','G193','ASIA','AUNZ','EAST','NAMER','NATL_AR2','NHEM','NPO','SHEM','TROPICS'
                     ],
                     'var_dict': {
                         'ICESEV': {'fcst_var_names': ['ICESEV'],
@@ -2102,7 +2102,7 @@ class Reference():
                     'plot_stats_list': 'bias, rmse, fbar_obar',
                     'interp': 'NEAREST',
                     'vx_mask_list' : [
-                        'G45','G193','ASIA','AUNZ','EAST','NAMER','NATL_AR2','NHEM','NPO','SHEM','TROPICS'
+                        'G045','G193','ASIA','AUNZ','EAST','NAMER','NATL_AR2','NHEM','NPO','SHEM','TROPICS'
                     ],
                     'var_dict': {
                         'TMP': {'fcst_var_names': ['TMP'],
